+++ Tiles/DriverConsole/Tiles/Bodies/Injuries/IInjuryReportCalc.cs
@@ -14,7 +14,7 @@ namespace Tiles.Bodies.Injuries
     {
         IInjuryReport CalculateMaterialStrike(
             IEnumerable<IItem> armorItems, StressMode stressMode,
-            double momentum, double contactArea, int maxPenetration,
+            double momentum, double contactArea, double maxPenetration,
             IBody targetBody, IBodyPart targetPart, IMaterial strikerMat, double sharpness);
     }
 }
